feat: infrastructure hardening, diagnostic instrumentation, and swarm manifest synchronization
- Hardened start.bat to force builds and clean stale processes
- Added high-visibility red-banner diagnostics to main.rs
- Implemented println! fallbacks in Config::load to bypass tracing order-of-ops
- Synchronized latest agent manifests and core infrastructure changes
- Resolved log clutter (skills directory mismatch)
- Hardened .gitignore to protect local data

+++ crates/agent/src/orchestration/continuation.rs
@@ -146,7 +146,7 @@ impl ContinuationEngine {
 
         let actual_delay = if self.config.use_exponential_backoff {
             // Exponential backoff: delay = min(base * 2^(n-1), max)
-            let backoff = self.config.default_delay_ms * 2u64.pow((*count - 1) as u32);
+            let backoff = self.config.default_delay_ms * 2u64.pow(*count - 1);
             backoff.min(self.config.max_delay_ms).min(delay_ms * 2)
         } else {
             delay_ms
